Seperate Fragments
Seperate several fragment from SetupActivity class. All the new fragments class contained in the package "fragments".
